ocr-reconstruct/requirements.txt
opencv-python-headless
pytesseract
numpy
pillow
matplotlib
pytest
scikit-image
scipy
tqdm
